Refactor client side scripts, use common js pattern
