+++ app/main.js
@@ -1,0 +1,6 @@
+var PostBox = require('./components/post/PostBox');
+
+React.render(
+    <PostBox url="/api/posts" />,
+    document.getElementById('content')
+);
