+++ app/utilities/strip.js
@@ -1,0 +1,6 @@
+module.exports = function(html) {
+    var tmp = document.createElement("DIV");
+    tmp.innerHTML = html;
+
+    return tmp.textContent || tmp.innerText || "";
+};
